changes to file structure and basic layout
