+++ todoredux/src/components/Button/Button.css
@@ -1,0 +1,4 @@
+button {
+    width: auto;
+    font-weight: bold;
+}
