+++ todoredux/src/components/Form/Form.css
@@ -1,0 +1,3 @@
+input {
+    width: 40%; 
+}
