+++ todoredux/src/components/Form/Form.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import Button from '../../components/Button/Button.js';
+import './Form.css';
 
 class Form extends Component {
     render () {
@@ -7,8 +8,6 @@ class Form extends Component {
             <form>
               <input name="ToDoInput" value=''></input>
               <Button name='Add' />
-              <Button name='Edit' />
-              <Button name='Delete' />  
             </form> 
         );
     }
