+++ todoredux/src/components/List/List.js
@@ -1,14 +1,16 @@
 import React, { Component } from 'react';
+import ListItem from '../ListItem/ListItem';
 
-const bttnsArr = {...this.props};
-console.log(bttnsArr);
 
 class List extends Component {
     render () {
+        const toDoItems = this.props.toDoList.map((item, index) => {
+            return <ListItem key={index} toDoText={item}/>
+        });
         return (
-            <ul>
-                <li>test</li>
-            </ul>
+            <div>
+           {toDoItems}
+           </div>
         );
     }
 }
