+++ todoredux/src/components/ListItem/ListItem.css
@@ -1,0 +1,8 @@
+.list-item-container {
+    display: flex;
+    flex-direction: row;
+}
+
+#todo {
+align-items: center;
+}
